pyproject.toml
[project]
name = "agentify-example-tau-bench"
version = "0.1.0"
description = "Example code of agentifying tau-bench for the blog `Agentify the Agent Assessment`."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.8",
    "agentbeats>=1.2.6",
    "dotenv>=0.9.9",
    "tau-bench",
    "typer>=0.19.2",
    "uvicorn>=0.37.0",
]

[tool.uv.sources]
tau-bench = { git = "https://github.com/sierra-research/tau-bench.git" }
